web_app/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.2=py39hecd8cb5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39hca72f7f_0
asgiref=3.4.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.2=py39he3068b8_1
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py39h9ed2024_1003
ca-certificates=2022.3.29=hecd8cb5_0
certifi=2021.10.8=py39hecd8cb5_2
cffi=1.15.0=py39hc55c11b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=36.0.0=py39hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py39he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
django=4.0.2=py39h6e9494a_0
django-pandas=0.6.6=pyhd8ed1ab_0
entrypoints=0.3=py39hecd8cb5_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
giflib=5.2.1=haf1e3a3_0
icu=58.2=h0a44026_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.8.2=py39hecd8cb5_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.4.1=py39hecd8cb5_1
ipython=7.31.1=py39hecd8cb5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py39hecd8cb5_1
jinja2=3.0.2=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h9ed2024_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py39hecd8cb5_7
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py39hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.2=py39he9d5cce_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.2.0=h87d7836_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.3=h23ab428_1
markupsafe=2.0.1=py39h9ed2024_0
matplotlib=3.5.0=py39hecd8cb5_0
matplotlib-base=3.5.0=py39h4f681db_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h9ed2024_1000
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
munkres=1.1.4=py_0
nbclient=0.5.11=pyhd3eb1b0_0
nbconvert=6.3.0=py39hecd8cb5_0
nbformat=5.2.0=pyhd8ed1ab_0
ncurses=6.3=hca72f7f_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.8=py39hecd8cb5_0
numexpr=2.8.1=py39h2e5f0a9_0
numpy=1.21.2=py39h4b4dc7a_0
numpy-base=1.21.2=py39he0bd621_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1n=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py39h743cdd8_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py39h98e4679_0
pip=21.2.4=py39hecd8cb5_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h23ab428_6
pyrsistent=0.18.0=py39hca72f7f_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.7=h88f2d9e_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd3eb1b0_0
pyzmq=22.3.0=py39he9d5cce_2
qt=5.9.7=h468cd18_1
qtconsole=5.2.2=pyhd3eb1b0_0
qtpy=1.11.2=pyhd3eb1b0_0
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.0.2=py39hae1ba45_1
scipy=1.7.3=py39h8c7af03_0
seaborn=0.11.0=py_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py39hecd8cb5_0
sip=4.19.13=py39h23ab428_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=h707629a_0
sqlparse=0.4.1=py_0
terminado=0.13.1=py39hecd8cb5_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h7bc2e8c_0
tornado=6.1=py39h9ed2024_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.8=pyhd3eb1b0_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39hecd8cb5_1
websocket-client=1.2.3=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39hecd8cb5_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.4=h23ab428_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h4dc903c_4
zstd=1.4.9=h322a384_0
